+++ LipidCreator/Tutorial.cs
@@ -1686,8 +1686,8 @@ namespace LipidCreator
                     
                     // set MS2 data from tutorial two
                     ElementDictionary newElements = MS2Fragment.createEmptyElementDict();
-                    newElements[MS2Fragment.ELEMENT_POSITIONS["H"]] = 3;
-                    newElements[MS2Fragment.ELEMENT_POSITIONS["O"]] = 2;
+                    newElements[(int)MS2Fragment.ELEMENT_POSITIONS["H"]] = 3;
+                    newElements[(int)MS2Fragment.ELEMENT_POSITIONS["O"]] = 2;
                     MS2Fragment newFragment = new MS2Fragment("testFrag", "testFrag", Lipid.chargeToAdduct[1], "", newElements, "FA1");
                     newFragment.userDefined = true;
                     creatorGUI.lipidCreator.allFragments["PG"][true]["testFrag"] = newFragment;
